more scalable command structure
