+++ src/cli/commands/index.ts
@@ -1,0 +1,2 @@
+export * from "./deposit";
+export * from "./eth1-private-network";
