+++ src/cli/commands/interface.ts
@@ -1,0 +1,7 @@
+import {CommanderStatic} from "commander";
+
+export interface ICliCommand {
+
+  register(commander: CommanderStatic): void;
+
+}
